perf(web): avoid full doc scan on footer cursor moves
Extract heading parsing utilities and only rebuild the outline when the document changes, not on every cursor move.
